Renaming stencil folder to mustache for clarity
